fix: build
fix: separate cjs and esm targets
fix: remove vite browser build
