+++ package.json
@@ -2,23 +2,22 @@
   "name": "@solarpunkltd/file-manager-lib",
   "version": "1.0.4",
   "description": "A file manager for storing and handling data on Swarm.",
-  "main": "dist/index.d.ts",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
   "types": "dist/types/index.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/types/index.d.ts",
-      "require": "./dist/index.d.ts",
-      "import": "./dist/index.d.ts",
-      "default": "./dist/index.d.ts"
+      "require": "./dist/cjs/index.js",
+      "import": "./dist/esm/index.js"
     },
     "./package.json": "./package.json"
   },
   "scripts": {
     "prepublishOnly": "cross-env NODE_ENV=production npm run build",
-    "build": "rimraf dist && npm run build:node && npm run build:types && npm run build:browser",
-    "build:node": "tsc -p tsconfig.json",
+    "build": "rimraf dist && npm run build:types && npm run build:cjs && npm run build:esm",
+    "build:cjs": "tsc -p tsconfig.cjs.json",
+    "build:esm": "tsc -p tsconfig.esm.json",
     "build:types": "tsc --emitDeclarationOnly --declaration --outDir dist/types",
-    "build:browser": "vite build",
     "test": "jest --config=jest.config.ts --runInBand --verbose --silent",
     "test:keep": "cross-env KEEP_BEE_DIRS=true npm run test",
     "test:ut": "cross-env KEEP_BEE_DIRS=true npm run test -- --selectProjects=unit",
@@ -31,12 +30,6 @@
   "files": [
     "dist"
   ],
-  "browser": {
-    "fs": false,
-    "path": false,
-    "stream": false,
-    "crypto": false
-  },
   "keywords": [
     "swarm",
     "bee",
@@ -51,47 +44,45 @@
   "homepage": "https://github.com/Solar-Punk-Ltd/file-manager-lib",
   "license": "Apache-2.0",
   "dependencies": {
-    "@ethersphere/bee-js": "^10.2.0",
-    "cafe-utility": "^33.3.2"
+    "@ethersphere/bee-js": "^10.4.0",
+    "cafe-utility": "^33.3.5"
   },
   "devDependencies": {
-    "@babel/cli": "^7.19.3",
-    "@babel/core": "^7.26.10",
-    "@babel/plugin-transform-class-properties": "^7.25.9",
-    "@babel/plugin-transform-runtime": "^7.17.0",
-    "@babel/preset-env": "^7.26.9",
-    "@babel/preset-typescript": "^7.26.0",
-    "@eslint/js": "^9.22.0",
-    "@jest/globals": "^29.7.0",
+    "@babel/cli": "^7.28.6",
+    "@babel/core": "^7.28.6",
+    "@babel/plugin-transform-class-properties": "^7.28.6",
+    "@babel/plugin-transform-runtime": "^7.28.5",
+    "@babel/preset-env": "^7.28.6",
+    "@babel/preset-typescript": "^7.28.5",
+    "@eslint/js": "^9.39.2",
+    "@jest/globals": "^30.2.0",
     "@types/event-emitter": "^0.3.5",
     "@types/fs-extra": "^11.0.4",
-    "@types/jest": "^29.5.14",
-    "@types/node": "^22.7.5",
-    "@typescript-eslint/eslint-plugin": "^8.26.1",
-    "@typescript-eslint/parser": "^8.26.1",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^25.1.0",
+    "@typescript-eslint/eslint-plugin": "^8.54.0",
+    "@typescript-eslint/parser": "^8.54.0",
     "babel-loader": "^10.0.0",
     "babel-plugin-add-import-extension": "^1.6.0",
-    "cross-env": "^7.0.3",
-    "eslint": "^9.22.0",
-    "eslint-config-prettier": "^10.1.1",
-    "eslint-import-resolver-typescript": "^4.2.0",
-    "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-jest": "^28.11.0",
+    "cross-env": "^10.1.0",
+    "eslint": "^9.39.2",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-import-resolver-typescript": "^4.4.4",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-jest": "^29.12.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^5.2.3",
-    "eslint-plugin-react-refresh": "^0.4.5",
+    "eslint-plugin-prettier": "^5.5.5",
+    "eslint-plugin-react-refresh": "^0.4.26",
     "eslint-plugin-simple-import-sort": "^12.1.1",
-    "jest": "^29.7.0",
-    "jest-cli": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "prettier": "^3.5.3",
-    "rimraf": "^6.0.1",
-    "std-env": "^3.8.1",
-    "ts-jest": "^29.2.5",
+    "jest": "^30.2.0",
+    "jest-cli": "^30.2.0",
+    "jest-environment-jsdom": "^30.2.0",
+    "prettier": "^3.8.1",
+    "rimraf": "^6.1.2",
+    "std-env": "^3.10.0",
+    "ts-jest": "^29.4.6",
     "ts-node": "^10.9.2",
-    "typescript": "^5.2.2",
-    "vite": "^6.2.2",
-    "vite-plugin-dts": "^4.5.3"
+    "typescript": "^5.9.3"
   },
   "engines": {
     "node": ">=14"
