+++ src/upload/index.ts
@@ -1,17 +1,14 @@
-import {
+import { isNode } from 'std-env';
+import type {
   Bee,
   BeeRequestOptions,
   CollectionUploadOptions,
   FileUploadOptions,
   RedundantUploadOptions,
-  UploadResult,
 } from '@ethersphere/bee-js';
-import { DriveInfo, FileInfoOptions, BrowserUploadOptions, NodeUploadOptions } from '../utils';
-import { ReferenceWithHistory } from '../utils/types';
-import { isNode } from 'std-env';
-import { uploadNode } from './upload.node';
-import { uploadBrowser } from './upload.browser';
-import { FileInfoError } from '../utils/errors';
+import type { DriveInfo, FileInfoOptions } from '../utils';
+import type { ReferenceWithHistory } from '../utils/types';
+import { processUploadBrowser } from './upload.browser';
 
 export async function processUpload(
   bee: Bee,
@@ -20,44 +17,10 @@ export async function processUpload(
   uploadOptions?: RedundantUploadOptions | FileUploadOptions | CollectionUploadOptions,
   requestOptions?: BeeRequestOptions,
 ): Promise<ReferenceWithHistory> {
-  uploadOptions = { ...uploadOptions, redundancyLevel: driveInfo.redundancyLevel };
-
-  if (fileOptions.file) {
-    return {
-      reference: fileOptions.file.reference.toString(),
-      historyRef: fileOptions.file.historyRef.toString(),
-    } as ReferenceWithHistory;
-  }
-
-  const batchId = driveInfo.batchId;
-  let uploadResult: UploadResult;
-
   if (isNode) {
-    const nodeOptions: NodeUploadOptions = fileOptions as NodeUploadOptions;
-
-    if (!nodeOptions.path) {
-      throw new FileInfoError('File path is required.');
-    }
-
-    uploadResult = await uploadNode(bee, batchId, nodeOptions, uploadOptions, requestOptions);
-  } else {
-    const browserOptions: BrowserUploadOptions = fileOptions as BrowserUploadOptions;
-
-    if (!browserOptions.files) {
-      throw new FileInfoError('Files are required.');
-    }
-
-    uploadResult = await uploadBrowser(
-      bee,
-      batchId,
-      browserOptions,
-      uploadOptions as RedundantUploadOptions,
-      requestOptions,
-    );
+    const { processUploadNode } = await import('./upload.node');
+    return processUploadNode(bee, driveInfo, fileOptions, uploadOptions, requestOptions);
   }
 
-  return {
-    reference: uploadResult.reference.toString(),
-    historyRef: uploadResult.historyAddress.getOrThrow().toString(),
-  } as ReferenceWithHistory;
+  return processUploadBrowser(bee, driveInfo, fileOptions, uploadOptions, requestOptions);
 }
