+++ src/upload/upload.node.ts
@@ -4,14 +4,20 @@ import {
   BeeRequestOptions,
   CollectionUploadOptions,
   FileUploadOptions,
+  RedundantUploadOptions,
   UploadResult,
 } from '@ethersphere/bee-js';
 
 import { FileError } from '../utils/errors';
-import { isDir, readFile } from '../utils/node';
-import { NodeUploadOptions, WrappedUploadResult } from '../utils/types';
+import {
+  DriveInfo,
+  FileInfoOptions,
+  NodeUploadOptions,
+  ReferenceWithHistory,
+  WrappedUploadResult,
+} from '../utils/types';
 
-export async function uploadNode(
+async function uploadNode(
   bee: Bee,
   batchId: string | BatchId,
   nodeOptions: NodeUploadOptions,
@@ -53,11 +59,14 @@ async function uploadFileOrDirectory(
   uploadOptions?: CollectionUploadOptions | FileUploadOptions,
   requestOptions?: BeeRequestOptions,
 ): Promise<UploadResult> {
-  if (isDir(resolvedPath)) {
+  const { isDir } = await import('../utils/fs/fs.node');
+  const isPathDir = await isDir(resolvedPath);
+
+  if (isPathDir) {
     return uploadDirectory(bee, batchId, resolvedPath, uploadOptions as CollectionUploadOptions, requestOptions);
-  } else {
-    return uploadFile(bee, batchId, resolvedPath, uploadOptions as FileUploadOptions, requestOptions);
   }
+
+  return uploadFile(bee, batchId, resolvedPath, uploadOptions as FileUploadOptions, requestOptions);
 }
 
 async function uploadFile(
@@ -68,7 +77,8 @@ async function uploadFile(
   requestOptions?: BeeRequestOptions,
 ): Promise<UploadResult> {
   try {
-    const { data, name, contentType } = readFile(resolvedPath);
+    const { readFile } = await import('../utils/fs/fs.node');
+    const { data, name, contentType } = await readFile(resolvedPath);
 
     return await bee.uploadFile(
       batchId,
@@ -97,4 +107,36 @@ async function uploadDirectory(
   } catch (error: any) {
     throw new FileError(`Failed to upload directory ${resolvedPath}: ${error}`);
   }
+}
+
+export async function processUploadNode(
+  bee: Bee,
+  driveInfo: DriveInfo,
+  fileOptions: FileInfoOptions,
+  uploadOptions?: RedundantUploadOptions | FileUploadOptions | CollectionUploadOptions,
+  requestOptions?: BeeRequestOptions,
+): Promise<ReferenceWithHistory> {
+  uploadOptions = { ...uploadOptions, redundancyLevel: driveInfo.redundancyLevel };
+
+  if (fileOptions.file) {
+    return {
+      reference: fileOptions.file.reference.toString(),
+      historyRef: fileOptions.file.historyRef.toString(),
+    } as ReferenceWithHistory;
+  }
+
+  const batchId = driveInfo.batchId;
+
+  const nodeOptions: NodeUploadOptions = fileOptions as NodeUploadOptions;
+
+  if (!nodeOptions.path) {
+    throw new Error('File path is required.');
+  }
+
+  const uploadResult = await uploadNode(bee, batchId, nodeOptions, uploadOptions, requestOptions);
+
+  return {
+    reference: uploadResult.reference.toString(),
+    historyRef: uploadResult.historyAddress.getOrThrow().toString(),
+  } as ReferenceWithHistory;
 }
