+++ tsconfig.esm.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["src"],
+  "compilerOptions": {
+    "module": "ESNext",
+    "outDir": "dist/esm",
+    "rootDir": "src"
+  }
+}
