+++ tsconfig.json
@@ -2,18 +2,19 @@
   "include": ["src", "tests"],
   "exclude": ["node_modules", "dist"],
   "compilerOptions": {
-    "target": "ES2020",
-    "lib": ["ES2020", "DOM", "DOM.Iterable"],
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
+    "target": "ES2022",
+    "lib": ["ES2022", "DOM", "DOM.Iterable"],
+    "module": "ES2022",
+    "baseUrl": "./src",
+    "moduleResolution": "node",
     "resolveJsonModule": true,
+    "declaration": true,
+    "emitDeclarationOnly": false,
+    "esModuleInterop": true,
     "types": ["node", "jest"],
     "skipLibCheck": true,
-    "declaration": true,
     "declarationMap": true,
-    "declarationDir": "./dist/types",
     "sourceMap": true,
-    "outDir": "dist",
     "strict": true,
     "alwaysStrict": true,
     "noUnusedLocals": true,
@@ -22,7 +23,6 @@
     "strictPropertyInitialization": true,
     "isolatedModules": true,
     "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
     "noImplicitAny": true,
     "typeRoots": ["src/@types", "node_modules/@types"],
     "noImplicitReturns": true
